add: ディレクトリ構造を変更
