+++ lib/ui/features/movie/movie_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class MovieScreen extends StatelessWidget {
-  const MovieScreen({Key? key}) : super(key: key);
+class MoviePage extends StatelessWidget {
+  const MoviePage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
