+++ lib/ui/features/shop/shop_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-class ShopScreen extends StatelessWidget {
-  const ShopScreen({Key? key}) : super(key: key);
+class ShopPage extends StatelessWidget {
+  const ShopPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
